backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
psycopg[binary]>=3.2,<3.3
